code/requirements.txt
colorama==0.3.9
coverage==4.5
Django==2.0.2
django-cors-headers==2.1.0
django-nose==1.4.5
djangorestframework==3.7.7
nose==1.3.7
pinocchio==0.4.2
psycopg2-binary==2.7.4
pytz==2018.3
